% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createData.R
\name{createData}
\alias{createData}
\title{Function used to generate the example data.}
\usage{
createData(rawDataPartsDir)
}
\arguments{
\item{rawDataPartsDir}{Directory containing the data to be distributed with the package.}
}
\description{
Function used to generate the example data.
}
