% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalls.R
\name{plot.markerResult}
\alias{plot.markerResult}
\title{Plot a manually called marker}
\usage{
plot.markerResult(x, allFounderNames, ...)
}
\arguments{
\item{x}{An object of class markerResult.}

\item{allFounderNames}{Optional list of the names of the founder lines replicates within the genetic data.}
}
\description{
Plot the results of manually calling a marker
}
\details{
This function plots the results of calling a marker manually using interactiveCall. In these cases a marker might have been called using 
the hierarchical Bayesia approach, or using DBSCAN.
}
